+++ garnet3d/main/src/priv/gfx/gpu/d3d11/d3d11ContextMgr.cpp
@@ -284,8 +284,8 @@ inline bool GN::gfx::D3D11Gpu::bindContextShader(
         GN_CASSERT( sizeof(AutoRef<Uniform>) == sizeof(Uniform*) );
 
         // apply GPU program resources
-        newProg->applyUniforms( (const Uniform * const *)newContext.uniforms.cptr(), newContext.uniforms.size(), skipDirtyCheck );
-        newProg->applyTextures( newContext.textures.cptr(), newContext.textures.size(), skipDirtyCheck );
+        newProg->applyUniforms( (const Uniform * const *)newContext.uniforms.cptr(), (uint32)newContext.uniforms.size(), skipDirtyCheck );
+        newProg->applyTextures( newContext.textures.cptr(), (uint32)newContext.textures.size(), skipDirtyCheck );
     }
     else if( skipDirtyCheck || (NULL != mContext.gpuProgram) )
     {
