+++ garnet3d/main/src/priv/gfx/util/gpu/mttexture.h
@@ -51,16 +51,16 @@ namespace GN { namespace gfx
         // ********************************
     public:
 
-        void   updateMipmap( size_t              face,
-                             size_t              level,
+        void   updateMipmap( uint32              face,
+                             uint32              level,
                              const Box<uint32> * area,
-                             size_t              rowPitch,
-                             size_t              slicePitch,
+                             uint32              rowPitch,
+                             uint32              slicePitch,
                              const void        * data,
                              SurfaceUpdateFlag   flag );
-        void   readMipmap( size_t face, size_t level, MipmapData & data );
-        void   blobWrite( const void * data, size_t length );
-        size_t blobRead( void * data );
+        void   readMipmap( uint32 face, uint32 level, MipmapData & data );
+        void   blobWrite( const void * data, uint32 length );
+        uint32 blobRead( void * data );
         void   generateMipmapPyramid();
         void * getAPIDependentData() const;
